Changed the manage tag dialog to suggest existing tags.
Changed the heading in the tag flow sankey view to be cut off when too long.

Corrected the hard coded text blocks of Legal Entity and Media Institution.

Changed the labeling of 'Set Legal Entity/Media Entity Tags' to 'Filter by Legal Entity/MediaEntity Tags'.

Fixed the issue of the application crashing when no columns for the tags are included are if the columns are not filled.

Fixed the issue of adding tags to each node that contains the relevant node name in its name.

Fixed the aggregated tag flow to show the sparkline barcharts.